Refactor bot to add twitter handler
